+++ thcrap_tasofro/src/th155_bmp_font.cpp
@@ -561,7 +561,7 @@ int patch_bmp_font(void *file_inout, size_t size_out, size_t size_in, const char
 	return 1;
 }
 
-extern "C" int BP_bmpfont_fix_parameters(x86_reg_t *regs, json_t *bp_info)
+extern "C" size_t BP_bmpfont_fix_parameters(x86_reg_t *regs, json_t *bp_info)
 {
 	// Parameters
 	// ----------
